requirements.txt
streamlit
langchain
google-generativeai
pinecone
python-dotenv
pandas